Interface/requirements.txt
# LangChain core libraries
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-text-splitters==0.0.1
langchain-google-genai==0.0.11
langsmith==0.1.17

# Google Generative AI
google-generativeai==0.4.1
google-auth>=2.27.0
google-api-python-client>=2.100.0

# Streamlit for web UI
streamlit==1.31.1

# HTTP requests
requests==2.32.3

# Environment variable management
python-dotenv==1.0.1

# Firestore integration
google-cloud-firestore


pandas>=2.0.0
bcrypt>=4.0.1
firebase-admin>=6.4.0
gunicorn==21.2.0
python-dotenv
requests>=2.32.0
